adjustments to DB, UI, REST and other stuff

+++ spaceMissionManager-persistance/src/main/java/cz/muni/fi/dao/SpacecraftDaoImpl.java
@@ -54,6 +54,9 @@ public class SpacecraftDaoImpl implements SpacecraftDao {
         if (del == null){
 	        throw new IllegalArgumentException("Spacecraft does not exist");
         }
+        if (del.getMission() != null && del.getMission().getSpacecrafts().size() == 1){
+            throw new IllegalArgumentException("Can't remove only spacecraft in mission");
+        }
         del.getComponents().forEach(p->del.removeComponent(p));
         em.remove(del);
         return spacecraft;
@@ -93,8 +96,16 @@ public class SpacecraftDaoImpl implements SpacecraftDao {
         if(spacecraft.getName() == null){
             throw new IllegalArgumentException("Spacecraft name must not be null");
         }
+        findSpacecraftById(spacecraft.getId()).getComponents().forEach(p->{
+            p.setSpacecraft(null);
+            em.merge(p);
+        });
         em.merge(spacecraft);
 
+        spacecraft.getComponents().forEach(p->{
+            p.setSpacecraft(spacecraft);
+            em.merge(p);
+        });
         em.flush();
         return spacecraft;
     }
